requirements.txt
# requirements.txt - 完整版
nonebot2[fastapi]>=2.0.0
nonebot-adapter-onebot>=2.0.0
aiohttp>=3.8.0
requests>=2.28.0
Pillow>=12.1.0
apscheduler>=3.10.0